+++ package.json
@@ -42,6 +42,9 @@
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
+    "@reduxjs/toolkit": "^2.8.2",
+    "@tanstack/react-query": "^5.85.5",
+    "axios": "^1.11.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "cmdk": "^0.2.1",
@@ -57,13 +60,17 @@
     "jspdf-autotable": "^3.8.1",
     "lodash.debounce": "^4.0.8",
     "lucide-react": "^0.323.0",
-    "next-themes": "^0.2.1",
     "next": "^14.2.5",
+    "next-themes": "^0.2.1",
     "notistack": "^3.0.0",
     "react": "^18.2.0",
     "react-day-picker": "^8.9.1",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^6.0.0",
+    "react-helmet-async": "^2.0.5",
     "react-hook-form": "^7.47.0",
+    "react-hot-toast": "^2.6.0",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^6.20.0",
     "recharts": "^2.8.0",
     "tailwind-merge": "^2.0.0",
