+++ src/components/AIServicesShowcase.jsx
