+++ src/components/BenefitsSection.jsx
