+++ src/components/CategoriesSection.jsx
