+++ src/components/GlobalPresenceSection.jsx
