+++ src/components/HowItWorksSection.jsx
