+++ src/components/InnovationResearchSection.jsx
