+++ src/components/PricingSection.jsx
