+++ src/components/TeamExpertiseSection.jsx
