+++ src/components/TechSolutionsSection.jsx
