+++ src/components/home/FeatureHighlights.jsx
