+++ src/context/auth/useAuthState.ts
@@ -1,0 +1,31 @@
+import { useState } from 'react';
+
+interface User {
+  id: string;
+  email: string;
+  name?: string;
+  avatar_url?: string;
+  created_at?: string;
+  updated_at?: string;
+}
+
+export const useAuthState = () => {
+  const [user, setUser] = useState<User | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [onboardingStep, setOnboardingStep] = useState(0);
+  const [tokens, setTokens] = useState<{ accessToken: string | null; refreshToken: string | null }>({
+    accessToken: null,
+    refreshToken: null
+  });
+
+  return {
+    user,
+    setUser,
+    isLoading,
+    setIsLoading,
+    onboardingStep,
+    setOnboardingStep,
+    tokens,
+    setTokens
+  };
+};
