+++ src/utils/authUtils.ts
@@ -1,0 +1,64 @@
+export const cleanupAuthState = () => {
+  // Clean up any stale authentication state
+  try {
+    // Clear any expired tokens or invalid auth data
+    const authData = localStorage.getItem('auth');
+    if (authData) {
+      const parsed = JSON.parse(authData);
+      if (parsed.expiresAt && new Date(parsed.expiresAt) < new Date()) {
+        localStorage.removeItem('auth');
+      }
+    }
+  } catch (error) {
+    console.error('Error cleaning up auth state:', error);
+    // Clear potentially corrupted data
+    localStorage.removeItem('auth');
+  }
+};
+
+export const isTokenValid = (token: string | null): boolean => {
+  if (!token) return false;
+  
+  try {
+    // Basic token validation - check if it's a JWT format
+    const parts = token.split('.');
+    if (parts.length !== 3) return false;
+    
+    // Check if token is expired (basic check)
+    const payload = JSON.parse(atob(parts[1]));
+    if (payload.exp && payload.exp * 1000 < Date.now()) {
+      return false;
+    }
+    
+    return true;
+  } catch (error) {
+    console.error('Error validating token:', error);
+    return false;
+  }
+};
+
+export const getStoredAuth = () => {
+  try {
+    const authData = localStorage.getItem('auth');
+    return authData ? JSON.parse(authData) : null;
+  } catch (error) {
+    console.error('Error getting stored auth:', error);
+    return null;
+  }
+};
+
+export const setStoredAuth = (authData: any) => {
+  try {
+    localStorage.setItem('auth', JSON.stringify(authData));
+  } catch (error) {
+    console.error('Error setting stored auth:', error);
+  }
+};
+
+export const clearStoredAuth = () => {
+  try {
+    localStorage.removeItem('auth');
+  } catch (error) {
+    console.error('Error clearing stored auth:', error);
+  }
+};
